api/pyproject.toml
[project]
name = "llm-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "boto3>=1.40.74",
    "chromadb>=1.3.4",
    "clerk-backend-api>=4.0.0",
    "crawl4ai>=0.4.0",
    "fastapi>=0.115.0",
    "langchain>=1.0.5",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=0.3.12",
    "langchain-postgres>=0.0.16",
    "openai>=2.7.1",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.2.3",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "crawlers*"]
namespaces = false
